deploy improvements

+++ deployer/src/emails.py
@@ -67,6 +67,7 @@ def _commit_push(config_name, action, config_dir):
     os.chdir(config_dir)
 
     with open(os.devnull, 'w') as dev_null:
+
         sp.call(['git', 'pull', '-r', 'origin', 'master'], stdout=dev_null, stderr=sp.STDOUT)
         sp.call(['git', 'add', txt], stdout=dev_null, stderr=sp.STDOUT)
         sp.call(['git', 'commit', '-m', commit_message], stdout=dev_null, stderr=sp.STDOUT)
@@ -77,7 +78,6 @@ def _commit_push(config_name, action, config_dir):
 
 def _write(emails, config_name, config_dir):
     file_path = path.join(config_dir, 'infos', config_name + '.json')
-    new_file = True
 
     obj = OrderedDict((
         ('name', config_name),
@@ -87,15 +87,12 @@ def _write(emails, config_name, config_dir):
     ))
 
     if path.isfile(file_path):
-        new_file = False
         with open(file_path, 'r') as f:
             obj = json.loads(f.read(), object_pairs_hook=OrderedDict)
             obj['emails'] = emails
 
     with open(file_path, 'w') as f:
         f.write(json.dumps(obj, separators=(',', ': '), indent=2))
-
-    return new_file
 
 
 def _prompt_emails(config_name, config_dir):
@@ -112,37 +109,17 @@ def _prompt_emails(config_name, config_dir):
 
 def add(config_name, config_dir, emails_to_add=None):
     if emails_to_add and len(emails_to_add) > 0:
-        new_file = _write(emails_to_add, config_name, config_dir)
-        add_emails(config_name, emails_to_add)
+        _write(emails_to_add, config_name, config_dir)
+
     else:
         emails = _prompt_emails(config_name, config_dir)
-        new_file = _write(emails, config_name, config_dir)
-        add_emails(config_name, emails)
+        _write(emails, config_name, config_dir)
 
-    if new_file:
-        _commit_push(config_name, 'Add', config_dir)
-    else:
-        _commit_push(config_name, 'Update', config_dir)
-
-
-def add_emails(config_name, emails):
-    from deployer.src.algolia_internal_api import add_user_to_index
-
-    for email in emails:
-        add_user_to_index(config_name, email)
-
-
-def delete_emails(config_name, emails):
-    from deployer.src.algolia_internal_api import remove_user_from_index
-
-    for email in emails:
-        remove_user_from_index(config_name, email)
+    _commit_push(config_name, 'Update', config_dir)
 
 
 def delete(config_name, config_dir):
     file_path = path.join(config_dir, 'infos', config_name + '.json')
-    if path.isfile(file_path):
-        emails = _retrieve(config_name, config_dir)
-        delete_emails(config_name, emails)
+    if path.isfile(file_path) and helpers.confirm('Delete emails for {}'.format(config_name)):
         os.remove(file_path)
         _commit_push(config_name, 'Delete', config_dir)
